fix(pre-open): continuous directional ranking score (baseline v2)
Replace the six-value raw_score lookup with a continuous 0–100 formula in
book pressure, delta_iev_ratio, and log-scaled intensity. Drop the unreachable
intensity>=1.0 confidence clamp. Pre-open ENTER/WATCH cutovers move onto
baseline-local enter_min_score/watch_min_score (62/48). Contract and PRE_OPEN
identity fork to v2; auction_quality remains cap-only.

+++ src/application/use_case/pre_open_workflow_use_case.py
@@ -138,7 +138,7 @@ class PreOpenSignalSummary:
     direction: str
     confidence: str
     auction_quality: str
-    raw_score: int
+    raw_score: float
     score: int
     strength: str
     entry_quality: str
